Moved entities to their appropriate modules
