+++ src/modules/session/entities/SessionParticipantRoles.ts
